# demos/sample-rest-stockapp/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api Documentation",
        "version": "1.0",
        "title": "Api Documentation",
        "termsOfService": "urn:tos",
        "contact": {},
        "license": { "name": "Apache 2.0", "url": "http://www.apache.org/licenses/LICENSE-2.0" }
    },
    "host": "localhost:8080",
    "basePath": "/",
...
    "paths": {
...
        "/v1/operate/{namespace}/{key}": {
            "post": {
                "tags": ["Operate operations"],
                "summary": "Perform multiple operations atomically on the specified record.",
                "operationId": "operateNamespaceKey",
                "consumes": ["application/json", "application/msgpack"],
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "name": "commitLevel",
                        "in": "query",
                        "description": "Desired consistency guarantee when committing a transaction on the server.",
                        "required": false,
                        "type": "string",
                        "default": "COMMIT_ALL",
                        "enum": ["COMMIT_ALL", "COMMIT_MASTER"]
                    },
                    {
                        "name": "consistencyLevel",
                        "in": "query",
                        "description": "How replicas should be consulted in a read operation to provide the desired consistency guarantee.",
                        "required": false,
                        "type": "string",
                        "default": "CONSISTENCY_ONE",
                        "enum": ["CONSISTENCY_ONE", "CONSISTENCY_ALL"]
                    },
                    {
                        "name": "durableDelete",
                        "in": "query",
                        "description": "If the transaction results in a record deletion, leave a tombstone for the record.",
                        "required": false,
                        "type": "boolean",
                        "default": false,
                        "x-example": false
                    },
                    {
                        "name": "expiration",
                        "in": "query",
                        "description": "Record expiration. Also known as ttl (time to live). Seconds record will live before being removed by the server.",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "generation",
                        "in": "query",
                        "description": "Expected generation. Generation is the number of times a record has been modified (including creation) on the server.",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "generationPolicy",
                        "in": "query",
                        "description": "Qualify how to handle record writes based on record generation.",
                        "required": false,
                        "type": "string",
                        "default": "NONE",
                        "enum": ["NONE", "EXPECT_GEN_EQUAL", "EXPECT_GEN_GT"]
                    },
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Userkey for the record.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "keytype",
                        "in": "query",
                        "description": "The Type of the userKey.",
                        "required": false,
                        "type": "string",
                        "default": "STRING",
                        "enum": ["STRING", "INTEGER", "BYTES", "DIGEST"]
                    },
                    {
                        "name": "linearizeRead",
                        "in": "query",
                        "description": "Force reads to be linearized for server namespaces that support strong consistency mode.",
                        "required": false,
                        "type": "boolean",
                        "default": false,
                        "x-example": false
                    },
                    {
                        "name": "namespace",
                        "in": "path",
                        "description": "Namespace for the record; equivalent to database name.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "operations",
                        "description": "An array of operation objects specifying the operations to perform on the record",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": { "$ref": "#/definitions/Operation" }
                        }
                    },
                    {
                        "name": "recordExistsAction",
                        "in": "query",
                        "description": "How to handle the existence of the record. This is ignored for POST/PUT/UPDATE kvs methods.",
                        "required": false,
                        "type": "string",
                        "enum": ["UPDATE", "UPDATE_ONLY", "REPLACE", "REPLACE_ONLY", "CREATE_ONLY"]
                    },
                    {
                        "name": "replica",
                        "in": "query",
                        "description": "Replica algorithm used to determine the target node for a single record command.",
                        "required": false,
                        "type": "string",
                        "default": "SEQUENCE",
                        "enum": ["MASTER", "MASTER_PROLES", "SEQUENCE", "RANDOM"]
                    },
                    {
                        "name": "sendKey",
                        "in": "query",
                        "description": "Send user defined key in addition to hash digest on both reads and writes.",
                        "required": false,
                        "type": "boolean",
                        "default": false,
                        "x-example": false
                    }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/Record" } },
                    "400": {
                        "description": "Invalid parameters or request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "403": {
                        "description": "Not authorized to access the resource",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "404": {
                        "description": "Namespace or record does not exist",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "409": {
                        "description": "Generation conflict",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            }
        },
        "/v1/operate/{namespace}/{set}/{key}": {
            "post": {
                "tags": ["Operate operations"],
                "summary": "Perform multiple operations atomically on the specified record.",
                "operationId": "operateNamespaceSetKey",
                "consumes": ["application/json", "application/msgpack"],
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "name": "commitLevel",
                        "in": "query",
                        "description": "Desired consistency guarantee when committing a transaction on the server.",
                        "required": false,
                        "type": "string",
                        "default": "COMMIT_ALL",
                        "enum": ["COMMIT_ALL", "COMMIT_MASTER"]
                    },
                    {
                        "name": "consistencyLevel",
                        "in": "query",
                        "description": "How replicas should be consulted in a read operation to provide the desired consistency guarantee.",
                        "required": false,
                        "type": "string",
                        "default": "CONSISTENCY_ONE",
                        "enum": ["CONSISTENCY_ONE", "CONSISTENCY_ALL"]
                    },
                    {
                        "name": "durableDelete",
                        "in": "query",
                        "description": "If the transaction results in a record deletion, leave a tombstone for the record.",
                        "required": false,
                        "type": "boolean",
                        "default": false,
                        "x-example": false
                    },
                    {
                        "name": "expiration",
                        "in": "query",
                        "description": "Record expiration. Also known as ttl (time to live). Seconds record will live before being removed by the server.",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "generation",
                        "in": "query",
                        "description": "Expected generation. Generation is the number of times a record has been modified (including creation) on the server.",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "generationPolicy",
                        "in": "query",
                        "description": "Qualify how to handle record writes based on record generation.",
                        "required": false,
                        "type": "string",
                        "default": "NONE",
                        "enum": ["NONE", "EXPECT_GEN_EQUAL", "EXPECT_GEN_GT"]
                    },
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Userkey for the record.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "keytype",
                        "in": "query",
                        "description": "The Type of the userKey.",
                        "required": false,
                        "type": "string",
                        "default": "STRING",
                        "enum": ["STRING", "INTEGER", "BYTES", "DIGEST"]
                    },
                    {
                        "name": "linearizeRead",
                        "in": "query",
                        "description": "Force reads to be linearized for server namespaces that support strong consistency mode.",
                        "required": false,
                        "type": "boolean",
                        "default": false,
                        "x-example": false
                    },
                    {
                        "name": "namespace",
                        "in": "path",
                        "description": "Namespace for the record; equivalent to database name.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "operations",
                        "description": "An array of operation objects specifying the operations to perform on the record",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": { "$ref": "#/definitions/Operation" }
                        }
                    },
                    {
                        "name": "recordExistsAction",
                        "in": "query",
                        "description": "How to handle the existence of the record. This is ignored for POST/PUT/UPDATE kvs methods.",
                        "required": false,
                        "type": "string",
                        "enum": ["UPDATE", "UPDATE_ONLY", "REPLACE", "REPLACE_ONLY", "CREATE_ONLY"]
                    },
                    {
                        "name": "replica",
                        "in": "query",
                        "description": "Replica algorithm used to determine the target node for a single record command.",
                        "required": false,
                        "type": "string",
                        "default": "SEQUENCE",
                        "enum": ["MASTER", "MASTER_PROLES", "SEQUENCE", "RANDOM"]
                    },
                    {
                        "name": "sendKey",
                        "in": "query",
                        "description": "Send user defined key in addition to hash digest on both reads and writes.",
                        "required": false,
                        "type": "boolean",
                        "default": false,
                        "x-example": false
                    },
                    {
                        "name": "set",
                        "in": "path",
                        "description": "Set for the record; equivalent to database table.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/Record" } },
                    "400": {
                        "description": "Invalid parameters or request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "403": {
                        "description": "Not authorized to access the resource",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "404": {
                        "description": "Namespace or record does not exist",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "409": {
                        "description": "Generation conflict",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
...
        "Operation": {
            "type": "object",
            "required": ["op_values", "operation"],
            "properties": { "op_values": { "type": "object" }, "operation": { "type": "string" } },
            "title": "Operation"
        },
...
        "Record": {
            "type": "object",
            "properties": {
                "bins": {
                    "type": "object",
                    "example": { "bin1": "val1", "pi": "3.14" },
                    "description": "A mapping from binName to binValue"
                },
                "generation": {
                    "type": "integer",
                    "format": "int32",
                    "example": 2,
                    "description": "The generation of the record."
                },
                "ttl": {
                    "type": "integer",
                    "format": "int32",
                    "example": 1000,
                    "description": "The time to live for the record, in seconds from now."
                }
            },
            "title": "Record"
        },
        "RestClientError": {
            "type": "object",
            "properties": {
                "inDoubt": {
                    "type": "boolean",
                    "example": false,
                    "description": "A boolean specifying whether it was possible that the operation succeeded. This is only included if true."
                },
                "message": {
                    "type": "string",
                    "example": "Error Message",
                    "description": "A message describing the cause of the error."
                }
            },
            "title": "RestClientError",
            "description": "Error object returned from unsuccesful operations."
        },
...
    }
}